working/gallag/requirements.txt
pygame>=2.6.1
dspy-ai>=3.1.3
openai>=2.24.0
python-dotenv>=1.0.1
streamlit>=1.30.0
